[FE-749] Add util to find the latest upload for doc viewer v2 (#13559)
## Describe your changes
PR [1/2] for adding a latest upload label to doc viewer v2. 

Context: it's nice for reviewers to know which upload is the latest. We
don't have a field denoting this yet, but the latest upload is just the
one with the highest `version` for that particular `identifier`. Eg in
`uploads = [{ version: 1, identifier: 'A', ...}, { version: 2,
identifier: 'A', ...}] `, the 2nd element is latest upload for that
identifier.

This PR: adds a util that adds an `isLatest` field to each upload in a
list of uploads. This will be used in the next PR inside the `select` of
the `use-documents` hook.

## Caveats


## Test Plan