[PR-396] make FE support null principal actor (#13665)
## Describe your changes

Update: totally reworked this! Previously a chance on `principal-actor`,
but seemed as if this didn't work so well, because we really just want
to not show the principal actor in that case. Instead, this PR gets rid
of the actor when we don't have it, and capitalizes the first letter of
the action. There's maybe a cleaner abstraction than this one, but I
think this does the job.

A huge part of this PR is tests and test configs! We are still using the
old types because complex types don't generate so well.


## Caveats


## Test Plan

Tests for `deactivate-org-role`, `update-org-role`, and
`create-org-role`.