Create dashboard facing API /entites/{fp_id}/onboardings (#13452)
## Describe your changes
In order to group entities on the dashboard by workflow or "playbook runs", we need to expose a list of workflows that a given user has gone through. This PR exposes an endpoint for `/entities/{fp_id}/onboardings`, to allow the dashboard to enumerate the list of playbooks that were run.

## Caveats
N/A

## Test Plan
* See integration tests