further refactor (#13692)
## Describe your changes
We have 2 ways of generating DIs from incode OCR. For Doc first playbook flows we have `doc_first_id_data` in `complete.rs` and then we also generate ODKs in `compute_ocr_data`.

It would be helpful if we had a canonical single method that extracted the superset of DIs we wanted, and then we can expose methods to select from them in some manner.

the reason this comes up is because in the case for flexcar where we're 1) doing doc first 2) vaulting ODKs and 3) sending things to samba, we do not want the sets of logic to drift. The exact motivation is that samba requires address components, and currently we only vault Full Address

## Caveats

## Test Plan