vault AH response (#13754)
## Describe your changes
In order to show the activity history in the dashboard, we vault the response under a OCR DI. This is not the long term way to do this, but we already have prior art for curp validation response.

in the dashboard we can return this seqno associated with a document, and we're able to then decrypt the response associated with the document. DIs here are really weird so I don't love this, ideally we'd have a special purpose API

In this case, we have `SambaOrder` which keeps track of `document_id` and we store the seqno of the DI on `SambaReport` therefore we can fetch the appropriate version for the frontend to decrypt in /hosted/documents

## Caveats

## Test Plan