Dashboard: awaiting bos event (#13774)
## Describe your changes

The PR is big because I added some configurations of storybook +
generated types on the same PR - My bad - I'll keep this for now though
as it should be simple to follow the code.

This is the first PR on the dashboard where we're using storybook +
interactive tests. In our dashboard, we have several audit trail events,
and right now it's very had to test because we need the backend. With
this approach, we can easily see the UI components for each event, and
even test. I'm happy to do a call explaining more

Here's a demo:


https://github.com/user-attachments/assets/69b4c244-74a3-40ed-9bd2-fc3e9f125870

Here's the interactive tests running:


https://github.com/user-attachments/assets/c5bb7e64-5f66-4c4e-bd88-a37a8931d85c


## Caveats


## Test Plan