[docs] Simplify recursive Schema rendering on API reference site (#13817)
## Describe your changes
The open API data model is recursive. Our old code was a kind of mess of
recursive rendering with duplicated logic at multiple different levels,
and it was only really designed to handle a few levels of nesting and
would bug out after deeper nesting.

In our old logic, the first level of schema was rendered slightly
differently than the >= second levels, not for any particular reason
other than implementation detail. This consolidates the logic to be
identical for all layers.

This also makes it much easier to update to support `anyOf`

## Caveats


## Test Plan
Here is a side-by-side that shows a lot of the improvements:
Old:

![image](https://github.com/user-attachments/assets/ff59903d-7f5c-44d5-a840-06a03b0d7903)

New:

![image](https://github.com/user-attachments/assets/9978769c-8fd7-4617-9692-d55f28e79a87)