[BE-1813] Support anyOf on our docs site (#13773)
## Describe your changes
In #12467, we started rendering `anyOf` in our open API schemas for
complex rust enum types. This starts to handle these `anyOf` types on
our docs site.

Right now, we only use `anyOf` types in our non-public docs site - just
for dashboard- and bifrost-facing APIs. So this display is mostly for
our own developers looking up the API docs of a dashboard/bifrost API.
(and it's okay that the display of these types is a little clunky - much
better than before where we had almost no documentation of types like
`UserTimeline` and `AuditTrail`)

## Caveats


## Test Plan
Locally:

https://github.com/user-attachments/assets/08dc61c6-d897-4e80-9e0b-e830d1f8b37c