[PR-399] Security logs --> Open API types (#13916)
## Describe your changes

Sorry for such a large PR! Many of the components on this page are quite
interconnected / dependent on one another, so was easiest to just make
one large PR that junks all of this old code. Also ran the OpenAPI
generation logic. All tests should pass and this should be a no-op to
the user.

## Caveats

## Test Plan

Existing tests, linter, and typecheck.