[BE-1787 update API keys audit event (#13839)
## Describe your changes

Title! This uses the new logic to get secondary data, too.

NOTE: You can update the name of an API key on the backend, but we don't
expose any functionality to this to users, so we don't have it in the
logs / I think don't need to log here - but we might want to make this
event more flexible? Curious for thoughts here

## Caveats


## Test Plan

`test_api_keys.py`