Only allow starting onboarding onto active/latest OBC (#13878)
## Describe your changes
New onboards are bootstrapped with a key. In-progress onboarding sessions are identified by an OBC ID. This PR updates the OBC/Playbook getters to clarify that a key doesn't uniquely identify an OBC. The effect of this is that new onboards use the latest configuration, and in-progress onboards are allowed to continue with an old configuration version. This is tested upstack once we add the API to create new OBC versions.

## Caveats

## Test Plan