Include context on where a database NotFound error originates (#13955)
## Describe your changes
We normally get no metadata on where the error originates for a NotFound
error. diesel can't tell us which table the query came from since the
query could have had a join. So this at least includes the caller's file
in the debug output of the error, which won't be visible via the prod
API but could then be looked up given the `support_id` of an error
response

## Caveats
I'm also going to try to include something like this by default into
`FpError`. I might include the full backtrace in `FpError`, which would
make this not really necessary. Going to play around a bit

## Test Plan
```
$ curl -s -X POST http://localhost:8000/users/fp_id_blah/vault/decrypt   -u sk_test_bO5lzm9r2HLlloOBeizIqmPM3tILO94n1i:  -d '{"fields": [], "reason": "blah"}' | jq
{
  "message": "Data not found",
  "code": null,
  "support_id": "45ef9448-5b53-4812-bf47-4fc84e376ec9",
  "debug": "DataNotFound(\"/Users/elliott/fp/monorepo/backend/components/db/core/src/models/scoped_vault.rs\")"
}
```