[BE-1743] Remove support for disabling fields in the KYC machine (#13942)
## Describe your changes
This feature was previously only used to disable editing the first name + last name + middle name when it was already entered for the primary beneficial owner, but there was a lot of unused code to support it in pretty much every field of the KYC machine.

There's not really a reason why the primary BO cannot update their name even though they already entered it earlier in the flow - so we stopped disabling the name fields in the KYC machine.

Now, we can kill this functionality to "disable" inputs entirely

## Caveats

## Test Plan