[BE-1818] Use allOf to represent additional fields on referenced schemas (#14030)
## Describe your changes
When we started using referenced schemas on the docs site, we lost a
little bit of functionality - since you couldn't provide any additional
metadata on top of a referenced schema, every place where we referenced
a schema used the same description, properties, extensions, etc. This
was problematic in a few places where we would want different
descriptions for two fields with the same type.

Bumps paperclip to support [this new
feature](https://github.com/onefootprint/paperclip/commit/0c2edcf56bae72b8bf6a006d8a4bfca05dc6c39a)
that generates a list of `allOf` when we have additional properties that
we want to add to a referenced schema.

## Caveats
This is a bit of an abuse of the open API spec i think. I think this is
usually used to add only additional properties to an existing schema,
whereas here in some cases we use it only to add a description to a
referenced schema. I added a little bit of custom logic to our
typescript type generation in `request-types` to properly handle our use
of `allOf`. Since we are the only consumers of our open API spec, this
is likely fine.

## Test Plan
A perfect example, the `POST /onboarding/session/validate` API response
has a `business` and `user` field, and each has a separate description.
We lost this metadata in a previous update, so it now displays like
this:

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/015c5e15-a5cd-46e3-a27c-f95ed9b5d9ce.png)
and here was the old open API spec:
```
            "ValidateResponse": {
                "type": "object",
                "properties": {
                    "business": {
                        "$ref": "#/components/schemas/EntityValidateResponse"
                    },
                    "user": {
                        "$ref": "#/components/schemas/EntityValidateResponse"
                    },
                    "user_auth": {
                        "$ref": "#/components/schemas/UserAuthResponse"
                    }
                },
                "required": [
                    "user_auth"
                ]
            },
```

### After this change
Here's how we display it on the docs site:


![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/67cbf8ca-5551-47a2-843d-adbb6e0d3b45.png)

And here's the new open API spec:
```
            "ValidateResponse": {
                "type": "object",
                "properties": {
                    "business": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/EntityValidateResponse"
                            },
                            {
                                "example": null,
                                "description": "Information on the business and its onboarding session. Provided after onboarding onto a KYB\n playbook.",
                                "type": "object"
                            }
                        ]
                    },
                    "user": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/EntityValidateResponse"
                            },
                            {
                                "description": "Information on the user and their onboarding session. Provided after onboarding onto a KYC\n or KYB playbook.",
                                "type": "object"
                            }
                        ]
                    },
                    "user_auth": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/UserAuthResponse"
                            },
                            {
                                "description": "Information on the authenticated user and the auth method they used.",
                                "type": "object"
                            }
                        ]
                    }
                },
                "required": [
                    "user_auth"
                ]
            },
```