[BE-1799] Support generating a link for users who end up in step up during POST /kyc (#13879)
## Describe your changes
In `POST /users/<>/kyc`, if a step-up rule was evaluated, we would
previously leave the user in an `incomplete` state with no ability for
the user to finish onboarding by providing the requested step-up
information.

Now, this adds an option to `POST /users/<>/kyc` (behind a feature gate)
to `generate_link_on_stepup`. When true, we'll create a new user auth
token linked to the exact workflow we just created. This link can be
used by the user to finish onboarding onto the playbook.

## Caveats
Weird things will happen with this link if the workflow ends up being
deactivated...

Are there any customers planning on using this yet? Or was this just to
be built for completeness?

## Test Plan
Integration test