Fix bug in business document workflow validation (#14087)
## Describe your changes

https://onefootprint.slack.com/archives/C05U1CAD6FQ/p1731973587943529?thread_ts=1731951865.298699&cid=C05U1CAD6FQ

Playbooks on document workflows strike again. Unfortunately, we require
every onboarding (aka workflow) to be associated with a playbook, even
though document workflows don't really read the playbook information
from the workflow. There was some validation when making a workflow for
a business that the playbook _must_ be a KYB playbook, even though it
doesn't actually matter for document workflows made for a business.

There was a very subtle case only evoked when (1) requesting a document
from a business and (2) the BO who fills out the form most recently
onboarded onto a KYC playbook. This bug has existed for a while but just
hasn't been evoked...

## Caveats


## Test Plan
Tested locally. Not going to add a test case for this since we will
hopefully migrate away from this soon