Add POST /org/playbook/{playbook_id}/restore (#14031)
## Describe your changes
This PR adds an endpoint that restores a playbook to a previous version.
It accomplishes this by creating a new version that has the same
configuration and rules as the old version. This yields a linear
history. There is currently no annotation that this version is a revert,
but I think audit logs can provide this functionality.

## Caveats
My brain is a little fuzzy at the moment from travel and I left this
half finished over the weekend so it's a little cold. I want to look at
this when I'm fresh tomorrow morning.

## Test Plan