Create new kind of token for async verification (#14151)
## Describe your changes
We'll soon start to support signup challenges using a new challenge
kind: `SmsLink`. These challenges will send a link to
`confirm.onefootprint.com` with a new `citok` in the URL hash. The citok
effectively acts as a the server-side state to track when a phone number
has been verified. citoks will _only_ ever be sent via SMS (ie never
returned in an HTTP response), so using them proves that an SMS was
received with the citok. They also wrap a user token, just to store some
context on the `user_vault_id` and `scoped_user_id` (I'm also happy to
just pass these identifiers directly in the token instead of indirectly
through another token, no strong opinion)

Notably, we create the `AuthEvent` for the identify session in this API.
This lets us get the InsightEvent for the user's mobile device. But,
most of the remaining logic to actually mark the phone as verified only
happens when the initial desktop client hits `POST
/hosted/identify/verify`.

## Caveats


## Test Plan
Will have integration tests in a future PR