[BE-1833] Support SmsLink challenge kind for login/signup challenges (#14153)
## Describe your changes
Implements the new `SmsLink` challenge kind behind a preview API. It can
also only be used on signup challenges. It will initially only be used
by the avis POS app: an agent on desktop will enter the user's phone
number to initiate an SmsLink challenge. After the user has clicked the
link, the agent on the avis POS app can proceed.

When selected, this will create a new `citok` and send an SMS containing
a link with this `citok` to the challenged phone number. This link will
hit an API that creates a new `AuthEvent` for the user's phone.

Only after this `citok` has been used to make an `AuthEvent` can the POS
desktop client proceed to hit `POST /hosted/identify/verify`, which will
mark the phone as verified and issue an auth token with proper
permissions

## Caveats
Eventually, we want to support a product where the agent in the POS app
can continue filing out user information _before_ the end user has
clicked the verification link. That will be a little bit more surgical,
so we'll omit for now

In a future PR, I'll also build a mechanism kind of like d2p that allows
the async email client to tell the POS client when to proceed. I don't
loooove the idea of the POS client just polling `POST
/hosted/identify/verify` and receiving HTTP 400s

## Test Plan