[BE-638] Deprecate honeycomb integration (#14179)
## Describe your changes
It's happening! We haven't really used honeycomb for the last few
months. Already downgraded our honeycomb plan to the free one, this just
stops emitting events

## Caveats
I'm not sure if the pulumi apply is going to be unhappy about deleting
the old honeycomb alerts that used to be managed in code... I can leave
them around if it is unhappy

## Test Plan