Deprecate ob_configuration tenant_id/is_live (#14186)
## Describe your changes
The original intention was to move reads to
`Playbook::tenant_id/is_live` so we could drop `OBC::tenant_id/is_live`.
However, the fact that it required so much rewiring made me realize for
DE v2 we probably want an abstraction above models that represents
playbooks in a way that's amenable to application code, rather than the
DB. DE v2 will probably take multiple tables to represent an OBC, and
passing around the individual rows as tuples isn't great for
readability/maintainability. I think we'll probably want to create a
`playbook` crate and use that to abstract models, especially as the data
structures become more complex (e.g. graphs, etc.).

A side effect of this work was a refactor to test fixtures that made a
few tests no longer depend on OBCs at all.

## Caveats


## Test Plan