Remove `allOf` in `ApiOnboardingRequirement` type (#14191)
## Describe your changes
The old serialization of onboarding requirements used a poorly supported
serde feature: `#[serde(flatten)]`. This was implemented as an `allOf`
in the open API spec.

Instead, let's just nest the whole requirement inline. This lets us get
rid of the `allOf`, making our open API spec a little more
client-language-agnostic.

## Caveats
This keeps the old inlined serialization around since a lot of our old
clients will keep reading this old field. It's probably going to be
extremely hard to tell when the client stops reading this old field...
It's possible some of our native SDKs are reading the old serialization,
so we can't drop the old serialization from the backend until all
clients are updated

## Test Plan