Fix flake in test_update_name, take 2 (#14241)
## Describe your changes
The flake is because we are changing a user name that is being read
multiple times by other tests. The first attempt
(https://github.com/onefootprint/monorepo/pull/14116) didn't work
because the user whose name we were changing was shared across multiple
tenants. This PR updates the test to change the name for a dedicated
user.

## Caveats


## Test Plan