[BE-1850] Copy playbook event (#14117)
## Describe your changes

Logic to copy playbooks. Basically:
- If we are changing tenants, we should create a `CopyPlaybook` event on
the source tenant, that refers to the source obc and the target tenant.
We should also create a `CreatePlaybook` event on the destination
tenant.
- If we are copying from the same tenant, we should only create a
`CreatePlaybook` event.

## Caveats

- In order to get the `tenant_name` without de-serializing, I use
`target_tenant_id` as metadata, create a way to fetch tenants in bulk
with secondary data, and handle this in the serializer
- These tests were failing before because we weren't disambiguating
between live and sandbox users. I've modified the
`assert_has_audit_event_with_details`, `list_audit_events_with_details`,
`get_audit_event_with_details` helpers with `is_live` parameters
- It was difficult to disambiguate new vs. old OBCs in `copy.rs` because
of the repetition of `obc`. I changed the variable names (which is why
the diff there is bigger) but don't feel super strongly here - if we
prefer the previous style, I am happy to change it back

## Test Plan

`test_ob_configs.py`