[BE-1874] manual review shows vault kind (#14244)
## Describe your changes

Tiny change! Realizing we need the `VaultKind` here because, if not,
there's no way to display "Manually reviewed this *user*" or "Manually
reviewed this *business*" in the frontend. (Also, to have the link go
correclty - i.e. /users/ vs. /businesses).

We already have this from the scoped vault, so we just need to serialize
out the kind! As a result we shouldn't need any backfill or anything /
this should be totally backwards compatible.

## Caveats


## Test Plan

Test the kind is properly serialized out in existing testing of
`test_actions.py`