[PR-451] remove firm employee gate from most security log events (#14226)
## Describe your changes

Title! Most of these no longer need to be behind an `isFirmEmployee`
gate! The only ones we are holding back are:
- all of the playbook events - seems better to release them all at once,
and there is still slightly more BE stuff I'm stuck on (first thing next
week!) we need for `CopyPlaybook` and `UpdatePlaybookStatus` and the
backfill (which - actually - may be delayed a bit untll we make some
other changes)
- `ManuallyReviewEntity` which I need to make some slight changes on 


Everything else should be good! I'll wait to ship this until monday/when
we can look at it.

## Caveats


## Test Plan


Testing throughout + in UI - these are super simple components and
mostly are handled by strict typing