Implement hosted in KMM OBC (#14280)
## Describe your changes
Broswer implementation is very different in iOS and Android. So, we are
not sharing browser code. This PR implements the android part of hosted
flow since it's more urgent. While all the OBC functions will be called
on `Footprint`, the `launchIdentify` and `handoff` will be called on
`FootprintHosted` because `Footprint` is common shared code and
`FootprintHosted` is android-specific. However, `FootprintHosted` uses
`Footprint` internally, `Footprint` must be initialized first.



https://github.com/user-attachments/assets/6762b9ae-c025-46ca-8e02-30ac26543d0c



## Caveats


## Test Plan