[BE-1882] Mark authed BO as immutable if workflow does not allow editing (#14345)
## Describe your changes
This is just a short-term fix.

There's currently a bug in the product where a user who onboards two
businesses onto the same playbook may try to edit their name on the BOs
form for the second business but cannot. This just marks the user as
immutable if the user's workflow is already complete.

## Caveats
The longer-term fix would actually create a new user workflow instead of
reusing the last one onboarding onto the same playbook. It's a little
bit more surgical though:
https://linear.app/footprint/issue/BE-1869/bugfix-workflow-guard-prevents-editing-name-on-beneficial-owner-screen


## Test Plan
Integration test.

Locally, verified that we disable the name when onboarding the second
business:
![Screenshot 2024-12-03 at
18.15.49.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/7302eb71-b369-4ed0-a1cc-b95de297c866.png)