Make more utils for cursor pagination (#14394)
## Describe your changes
Some cleanup to consolidate logic and make it very easy to add cursor
pagination.

Offset pagination has had a lot of these utils for a while, this just
modernizes cursor pagination.

The upstack PR shows an example of how simple it is to add cursor
pagination to an API after this PR

## Caveats


## Test Plan