[BE-1879] Paginate the `GET /entities/<>/timeline` API (#14304)
## Describe your changes
Once the frontend PR below this has been merged, the frontend will be
able to handle this more modern response.

## Caveats
This will break old versions of clients that are still loaded in the
browser, but they can simply refresh the page.

If we want to be tolerant of old clients, we could keep both the
paginated and non-paginated versions of the API around and have the
client explicitly request one or the other. But that's a little bit of
extra work, not sure if it's worth it. Lmk what you think


## Test Plan