feat(lists-v2):  [PR-481] table (#14381)
## Describe your changes

Title! Now we can actually see a table, there's more stuff to add back but this is closer!

![Screenshot 2024-12-04 at 2.33.19 PM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/24c25592-d351-46dc-bbfd-dec858b2c584.png)

## Caveats

Some of this functionality I need to add more screens in to build. Also, the types remain for the `PermissionGate`, but I can migrate this component while I do this / other related stuff.

## Test Plan