Use DataLifetimeSeqnoTxn for creating UserTimeline events (#14458)
## Describe your changes
We currently use seqnos to rebuild the state of a scoped vault at
different points in the timeline history. `UserTimeline` events have
associated seqnos, and we use those for the historical decrypt dashboard
feature. However, `UserTimeline` events were being created using seqnos
from `DataLifetime::get_current_seqno`, which does not require a locked
scoped vault, and therefore can not be used to create an ordering of
events within a scoped vault.

This problem surfaced via a flaky test, which attempted to assert that a
vault was empty at a given timeline event.

This PR adds a `DataLifetimeSeqnoTxn` to the arguments of
`UserTimeline::create` so seqnos can be well correlated. This will make
timeline seqnos more accurate, and in particular fix the flaky test,
since `data_collected` timeline events will have the same seqno as the
created data lifetimes.

This is not the only case of this class of bugs. We use
`DataLifetime::get_current_seqno` a lot of other places where the seqno
can't actually be used as an ordering or events. Ideally,
`DataLifetime::get_current_seqno` should only ever be used in read
paths. We can address this in future PRs.


## Caveats


## Test Plan