feat(lists-v2) [PR-472] actions page (#14427)
## Describe your changes

Actions dialog to delete a list.

NOTE: If we cancel the process, we need to refresh in order to get the confirmation dialog again. This is not behavior from this component, it seems to be how the confirmation dialog functions. I'm not sure if we want to keep it, but it seems alright for now.

<div class='graphite__hidden'>
          <div>🎥 Video uploaded on Graphite:</div>
            <a href="https://app.graphite.dev/media/video/XsTtzwNbP0GdQMHxY1hw/5a64013b-9fd7-4bae-b994-4ca20bb6ac87.mov">
              <img src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/XsTtzwNbP0GdQMHxY1hw/5a64013b-9fd7-4bae-b994-4ca20bb6ac87.mov">
            </a>
          </div>
<video src="https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/5a64013b-9fd7-4bae-b994-4ca20bb6ac87.mov">Screen Recording 2024-12-05 at 4.12.22 PM.mov</video>

## Caveats

## Test Plan