chore(migrate-api-keys): [FE-872] roles component (#14516)
## Describe your changes

Title! This was supposed to just be the tailwind change and the changing the type, but because the type is used in a bunch of form logic, I had to restructure how data is handled in several other components.

Note that `PermissionGate` has still not been updated to the new types — I've been waiting to fix this for a bit, seems like now is a good time.

Basically, the goal of this PR is: `roles.tsx` is on new types and tailwind (except for the `PermissionGate` stuff - I'll change all instances of this in the codebase in the next PR.

## Caveats

## Test Plan

Existing testing, typechecker, and I tested creating and editing roles in this section on this PR.