Relocate identify component to prepare for new signup flow (#14584)
## Describe your changes
In a follow-up PR, I'm going to implement a new version of the
server-driven signup flow, which will replace a portion of the legacy
identify flow.

This just moves the current identify flow to a new directory to prepare
for the new "signup" identify flow to be added. There are no logic
changes in this PR

## Caveats


## Test Plan