Implement new identify flow for signups (#14600)
## Describe your changes
In the PR below this, we add a new set of backend APIs that help to
better manage an "identify session." An identify session consists of the
whole process of collecting PII required by a playbook and verifying
login methods that are required by a playbook. The new set of APIs give
a server-driven set of requirements that can fulfill an identify
session, allowing us to kill a lot of hardcoded state machine
transitions on the frontend.

Notably, this _does not_ replace code used to _log into an existing
vault._ The new set of APIs (for now) only manage the _sign up_ process.

At a high level, here are the new APIs used by the signup flow:
More on each API, in roughly chronological order of when the client will
hit the APIs:
- `POST /hosted/identify/session`: initialize an identify session,
optionally with some vault data (like bootstrapped phone / email). This
returns an `itok` that is used for all of the below APIs
- `GET /hosted/identify/session/requirements`: fetch the set of
requirements needed to complete the identify session
- `PATCH /hosted/identify/session/vault`: vault data (usually just
phone/email) to the identify session. This actually writes vault data to
the unverified vault used by this identify session
- `POST /hosted/identify/session/challenge`: initiate a challenge of
either kind `sms` or `email` using the phone or email on file in this
identify session. This is practically only used for signup challenges
- `POST /hosted/identify/session/challenge/verify`: verify a challenge
with the received OTP
- `POST /hosted/identify/session/verify`: verify all identify
requirements are met and return a `utok`

More (outdated) context on the general identify v2 flow
[here](https://www.notion.so/onefootprint/Identify-Flow-v2-149eb271246680bc86f6ca8c4f2778e1).

## Caveats
This doesn't yet delete the old identify flow for sign ups - we have a
launch darkly tenant-based flag that opts single tenants into this new
signup flow.

The old signup code is deprecated in the WIP PR #14632 - feel free to
take a look for a little preview of what this replaces. After that PR,
the legacy `<Identify />` component really only handles logins.

## Test Plan
Will write some unit tests. E2e tests (after I merge the new set of
APIs) should all run against the new APIs

I did a lot of manual testing of really niche edge cases:
- "Back" button vs "Close" button logic, especially now that the back
buttons cause transitions between the new signup flow and the legacy
login flow
- Going back from an OTP challenge screen to a collect email / phone
screen and not clearing out the challenge data