Clarify IdentifyClient test library to avoid step_up confusion (#14721)
## Describe your changes
It's confusing to see `step_up()` in the `IdentifyClient`, when "step
up" means something different in the context of KYC. In this PR, I
clarify this by reworking the APIs to centralize on `login()` as the
only identify flow entrypoint for existing users. I move the test
assertions to the initializers, and conditionally run post-login
assertions whenever it's valid to do so. In practice, we never needed to
disable these post-login assertions. There was one case,
`test_replace_passkey`, in which the "stepped up" token returned from
`.login()` had the same scopes as the auth token used to identify the
user. To accommodate this, I relaxed the `_validate_login_token` check
to allow the new and old tokens to have the same scopes. I think this is
more correct in general.

## Caveats


## Test Plan