[PR-538] feat(footprint-wrapped): initial UI (#14780)
## Describe your changes

Created a new page `/wrapped` and add basic UI here. 

I figured it was better to do one PR with the initial UI changes before
we do all of the functionality stuff, that's coming:
- Like onboarding, we show users wrapped when they log in if they
haven't seen it before by saving a prop in the session
- Correctly handled redirects, i.e. when you open /wrapped it will
necessarily open this dialog but when you close it will redirect you
- Button in the user settings that will show you 2024 Footprint Wrapped

Dave did this without using OpenAPI typing / this isn't well-typed so I
did it manually.

![Screenshot 2024-12-20 at
10.14.34 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/2658b283-1cdc-455b-8085-5ebd1315d3a3.png)

![Screenshot 2024-12-20 at
10.14.13 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/XsTtzwNbP0GdQMHxY1hw/8c7f5021-c1e5-463b-8650-e9bb1376b5df.png)

## Caveats


## Test Plan

Dark mode and light mode on dev.