Update age libraries (#14720)
## Describe your changes
Pulls in the fix for
https://github.com/str4d/rage/security/advisories/GHSA-4fg7-vxc8-qx5w

Since we don't use the binary exec plugin functionality, we aren't
affected, but can't hurt to keep things updated. The bug is also rather
hard to exploit since:
> On UNIX systems, a directory matching age-plugin-* needs to exist in
the working directory for the attack to succeed.

~I don't think this needs a release of the `footprint-dr` CLI.~

This does change some behavior in the `footprint-dr` CLI, so I'll make a
patch release. The main difference is that we the `age` library no
longer lets us easily introspect identities to detect if they're a
yubikey. We could do this by looking for `AGE-PLUGIN-YUBIKEY` in the
identity file, but the introspection was only for the purpose of
silencing some irrelevant output for non-yubikey keys. In practice,
users will be using this tool mostly with yubikey keys, so I think it's
fine to print that output unconditionally to keep the crypto code
simple.

## Caveats


## Test Plan