Save sdk args token to DB even if deserialization fails (#14819)
## Describe your changes
We've moved a little bit more validation of SDK args into the JSON deserialization. Before this PR, if JSON deserialization failed, we would never save anythign to the DB and wouldn't have any record of what actually failed validation. This just starts to save an SDK args token regardless of whether the body is valid `SdkArgs`. But, we will continue to return an HTTP 400 to the client if the request is invalid

## Caveats

## Test Plan
```
# Returns an HTTP 400 when providing invalid SDK args data
$ curl http://localhost:8000/org/sdk_args -X POST -d '{"d": "flerpderp"}'
{"message":"missing field `kind`","code":null,"support_id":"feff23c6-b304-41ed-bb64-4fdd73d19d91","debug":"Error(\"missing field `kind`\", line: 0, column: 0)","location":"components/api/routes/org/src/sdk_args.rs:67:20"}

# Then, using the auth token hash from server logs:
$ curl -s http://localhost:8000/private/token/reveal -X POST -d '{"hash": "83b5cd8256664ac34ca80cdc2e67d0744a3598c5c87f7bab225b8faf0eb73244"}'  -H 'X-Fp-Protected-Custodian-Key: 2footprint' | jq
{
  "data": {
    "d": "flerpderp"
  },
  "kind": "sealed",
  "expires_at": "2024-12-23T21:24:21.647464Z",
  "token_hash": "83b5cd8256664ac34ca80cdc2e67d0744a3598c5c87f7bab225b8faf0eb73244"
}
```